requirements.txt
Flask
Flask-RESTX
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
Werkzeug
PyJWT
